+++ Assets/_Scripts/Core/Menus/GameplayMenuInstance.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+
+namespace Metroidvania.UI.Menus
+{
+    public abstract class GameplayMenuInstance : CanvasMenuBase
+    {
+        public GameplayMenuChannel channel { get; protected set; }
+
+        public abstract IEnumerator InitOperation(GameplayMenuChannel channel);
+
+        public abstract IEnumerator ReleaseOperation();
+    }
+}